new rev of GCC requires -pthread to compile.  This isn't as clean a solution as it could be since all code is being compiled with -pthread, rather than just the threaded apps.

+++ t/22_thread.c
@@ -83,7 +83,7 @@ int main(void)
             fail("pthread_join failed: %d\n", rc);
             return 13;
         }
-        pass("thread exited with status: %d\n", (int)t_rc);
+        pass("thread exited with status: %d\n", t_rc);
     }
 
     cmd = popen("cd tmp; sort * | uniq -d" , "r");
